New/requirements.txt
# SIGNUM SENTINEL — Python 3.12 (SPEC PART 9 + HuggingFace stack)
customtkinter>=5.2.0
tkinterdnd2>=0.3.0
psutil>=5.9.0
gputil>=1.4.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyexiftool>=0.5.6
tqdm>=4.66.0
safetensors>=0.4.0
huggingface_hub>=0.20.0
tokenizers>=0.15.0
sentencepiece>=0.1.99
transformers>=4.35.0
accelerate>=0.26.0
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
diffusers>=0.24.0
datasets>=2.16.0
peft>=0.7.0
pypdf>=3.0.0
pyinstaller>=6.0.0
pywin32>=306
